refactor: renamed files to be more informative
changed sentencepickers components into more informative names
