+++ frontend/src/pages/gameplay/components/fetchDaily.tsx
