+++ frontend/src/pages/gameplay/components/fetchNormal.tsx
