+++ frontend/src/pages/gameplay/dailychallengeScreen.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState } from "react";
 import { useAuth } from "../../auth/userContext";
 import { ColoredSentence } from "./components/coloredSentences";
-import { useDailySentencePicker } from "./components/dailyChallenge";
+import { useDailySentencePicker } from "./components/fetchDaily";
 import { useHandleInputChange } from "./components/handleInput";
 import { PostScore } from "./components/postScore";
